refactor: move single-consumer helpers from _helpers.ts into owning route modules
_helpers.ts shrank from ~445 to ~145 lines. Auth challenge DB ops,
Turnstile verification, CSV escaping, admin auth, leaderboard row
parsing, and display-name regex each moved to their sole consumer.
Only genuinely shared utilities (JSON responses, rate limiting,
wallet normalization, account provisioning) remain.
