+++ src/worker/routes/leaderboard.ts
@@ -1,12 +1,53 @@
-import { LEADERBOARD_LIMIT } from '../../domain/constants';
+import { clampTokenBalance, LEADERBOARD_LIMIT } from '../../domain/constants';
 import type { Env } from '../../types/worker-env';
 import { fetchAccountWithStats, shapeLeaderboardEntry } from '../accountRepo';
 import { getAuthenticatedAccountId } from '../session';
-import {
-  errorResponse,
-  jsonResponse,
-  parseLeaderboardEntryRow,
-} from './_helpers';
+import { errorResponse, jsonResponse } from './_helpers';
+
+function parseLeaderboardEntryRow(row: Record<string, unknown>): {
+  display_name: string | null;
+  token_balance: number;
+  leaderboard_eligible: number;
+  matches_played: number | null;
+  games_played: number | null;
+  coherent_games: number | null;
+  current_streak: number | null;
+  longest_streak: number | null;
+} | null {
+  const displayName = row.display_name;
+  if (!(displayName === null || typeof displayName === 'string')) return null;
+
+  const toNumber = (value: unknown): number | null => {
+    if (value === null || value === undefined) return null;
+    if (typeof value === 'number') return Number.isFinite(value) ? value : null;
+    if (typeof value === 'string' && value.length > 0) {
+      const parsed = Number(value);
+      return Number.isFinite(parsed) ? parsed : null;
+    }
+    return null;
+  };
+
+  const tokenBalance = toNumber(row.token_balance);
+  const leaderboardEligible = toNumber(row.leaderboard_eligible);
+  const matchesPlayed = toNumber(row.matches_played);
+  const gamesPlayed = toNumber(row.games_played);
+  const coherentGames = toNumber(row.coherent_games);
+  const currentStreak = toNumber(row.current_streak);
+  const longestStreak = toNumber(row.longest_streak);
+
+  if (tokenBalance === null || leaderboardEligible === null) return null;
+
+  return {
+    display_name: displayName,
+    token_balance: clampTokenBalance(tokenBalance),
+    leaderboard_eligible: leaderboardEligible,
+    matches_played: matchesPlayed,
+    games_played: gamesPlayed,
+    coherent_games: coherentGames,
+    current_streak: currentStreak,
+    longest_streak: longestStreak,
+  };
+}
 
 export async function handleLeaderboard(
   _request: Request,
